+++ src/main/java/com/distributore/dto/PurchaseResult.java
@@ -1,17 +1,16 @@
 package com.distributore.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record PurchaseResult(
 	    
 		boolean success,
-	    
 		String message,
-	    
 		ProductDto product,
-	    
-		Double change,
-	    
+		BigDecimal change,
 		LocalDateTime timestamp
-	
-		) {}
+		) 
+{
+
+}
